hwui: Relax the desiredPresent time in RenderAhead feature
During SW Vsync re-sync phase, the expectedPresent time slightly
exceeds desiredPresent time due to slight drift in HW Vsync samples
which leads to dropping of this buffer by BufferQueueConsumer. This
leads to buffer scarcity, as that buffer is meant to be consumed in
the next draw cycle. When RenderAhead feature is enabled, relax the
desiredPresent time by accounting Vsync period and App Vsync Offset
to avoid this buffer drop during the re-sync phase.

CRs-Fixed: 2551814
Change-Id: Ifd512c7b56c7e100970d39d67105cfb7b79b9fb9